LayoutTests/accessibility/aria-checked-mixed-value-expected.txt

Tests whether mixed values are reported properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Role: AXRole: AXRadioButton
Mixed: false


Role: AXRole: AXRadioMenuItem
Mixed: false


Role: AXRole: AXCheckBox
Mixed: true


Role: AXRole: AXCheckMenuItem
Mixed: true


Role: AXRole: AXCheckBox
Mixed: true


Role: AXRole: AXCheckBox
Mixed: false


PASS successfullyParsed is true

TEST COMPLETE
